requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
